+++ addOns/network/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.8.0] - 2023-05-03
 ### Added
 - Allow to log client side HTTP traffic for debug purposes, using the name `org.zaproxy.addon.network.http`.
 
@@ -94,6 +94,7 @@ would not be cleared when reusing the same message.
   - Options panel to manage the root CA certificate and issued certificates.
   - API endpoints to configure the validity of the root CA certificate and issued certificates ([Issue 4673](https://github.com/zaproxy/zaproxy/issues/4673)).
 
+[0.8.0]: https://github.com/zaproxy/zap-extensions/releases/network-v0.8.0
 [0.7.0]: https://github.com/zaproxy/zap-extensions/releases/network-v0.7.0
 [0.6.0]: https://github.com/zaproxy/zap-extensions/releases/network-v0.6.0
 [0.5.0]: https://github.com/zaproxy/zap-extensions/releases/network-v0.5.0